Move monitor_sfapi_job to sfapi_manager.py, add check for completion, do not override job state with
